add new task to update the vespa fields
