+++ backend/onyx/background/celery/tasks/user_file_processing/tasks.py
@@ -1,7 +1,10 @@
 import datetime
 import time
+from typing import Any
 from uuid import UUID
 
+import httpx
+import sqlalchemy as sa
 from celery import shared_task
 from celery import Task
 from redis.lock import Lock as RedisLock
@@ -10,6 +13,8 @@ from sqlalchemy import select
 from onyx.background.celery.apps.app_base import task_logger
 from onyx.background.celery.celery_utils import httpx_init_vespa_pool
 from onyx.background.celery.tasks.shared.RetryDocumentIndex import RetryDocumentIndex
+from onyx.background.celery.tasks.shared.tasks import LIGHT_SOFT_TIME_LIMIT
+from onyx.background.celery.tasks.shared.tasks import LIGHT_TIME_LIMIT
 from onyx.configs.app_configs import MANAGED_VESPA
 from onyx.configs.app_configs import VESPA_CLOUD_CERT_PATH
 from onyx.configs.app_configs import VESPA_CLOUD_KEY_PATH
@@ -28,6 +33,12 @@ from onyx.db.search_settings import get_active_search_settings
 from onyx.db.search_settings import get_active_search_settings_list
 from onyx.document_index.factory import get_default_document_index
 from onyx.document_index.interfaces import VespaDocumentUserFields
+from onyx.document_index.vespa.shared_utils.utils import get_vespa_http_client
+from onyx.document_index.vespa.shared_utils.utils import (
+    replace_invalid_doc_id_characters,
+)
+from onyx.document_index.vespa_constants import DOCUMENT_ID_ENDPOINT
+from onyx.document_index.vespa_constants import USER_PROJECT
 from onyx.httpx.httpx_pool import HttpxPool
 from onyx.indexing.adapters.user_file_indexing_adapter import UserFileIndexingAdapter
 from onyx.indexing.embedder import DefaultIndexingEmbedder
@@ -36,7 +47,6 @@ from onyx.natural_language_processing.search_nlp_models import (
     InformationContentClassificationModel,
 )
 from onyx.redis.redis_pool import get_redis_client
-from shared_configs.contextvars import CURRENT_TENANT_ID_CONTEXTVAR
 
 
 def _as_uuid(value: str | UUID) -> UUID:
@@ -64,7 +74,6 @@ def check_user_file_processing(self: Task, *, tenant_id: str) -> None:
     Uses direct Redis locks to avoid overlapping runs.
     """
     task_logger.info("check_user_file_processing - Starting")
-    CURRENT_TENANT_ID_CONTEXTVAR.set(tenant_id)
 
     redis_client = get_redis_client(tenant_id=tenant_id)
     lock: RedisLock = redis_client.lock(
@@ -116,7 +125,6 @@ def check_user_file_processing(self: Task, *, tenant_id: str) -> None:
 def process_single_user_file(self: Task, *, user_file_id: str, tenant_id: str) -> None:
     task_logger.info(f"process_single_user_file - Starting id={user_file_id}")
     start = time.monotonic()
-    CURRENT_TENANT_ID_CONTEXTVAR.set(tenant_id)
 
     redis_client = get_redis_client(tenant_id=tenant_id)
     file_lock: RedisLock = redis_client.lock(
@@ -272,7 +280,6 @@ def process_single_user_file(self: Task, *, user_file_id: str, tenant_id: str) -
 def check_for_user_file_project_sync(self: Task, *, tenant_id: str) -> None:
     """Scan for user files with PROJECT_SYNC status and enqueue per-file tasks."""
     task_logger.info("check_for_user_file_project_sync - Starting")
-    CURRENT_TENANT_ID_CONTEXTVAR.set(tenant_id)
 
     redis_client = get_redis_client(tenant_id=tenant_id)
     lock: RedisLock = redis_client.lock(
@@ -327,7 +334,6 @@ def process_single_user_file_project_sync(
     task_logger.info(
         f"process_single_user_file_project_sync - Starting id={user_file_id}"
     )
-    CURRENT_TENANT_ID_CONTEXTVAR.set(tenant_id)
 
     redis_client = get_redis_client(tenant_id=tenant_id)
     file_lock: RedisLock = redis_client.lock(
@@ -388,3 +394,179 @@ def process_single_user_file_project_sync(
             file_lock.release()
 
     return None
+
+
+def _normalize_legacy_user_file_doc_id(old_id: str) -> str:
+    # Convert USER_FILE_CONNECTOR__<uuid> -> FILE_CONNECTOR__<uuid> for legacy values
+    user_prefix = "USER_FILE_CONNECTOR__"
+    file_prefix = "FILE_CONNECTOR__"
+    if old_id.startswith(user_prefix):
+        remainder = old_id[len(user_prefix) :]
+        return file_prefix + remainder
+    return old_id
+
+
+def _visit_chunks(
+    *,
+    http_client: httpx.Client,
+    index_name: str,
+    selection: str,
+    continuation: str | None = None,
+) -> tuple[list[dict[str, Any]], str | None]:
+    base_url = DOCUMENT_ID_ENDPOINT.format(index_name=index_name)
+    params: dict[str, str] = {
+        "selection": selection,
+        "wantedDocumentCount": "1000",
+    }
+    if continuation:
+        params["continuation"] = continuation
+    resp = http_client.get(base_url, params=params, timeout=None)
+    resp.raise_for_status()
+    payload = resp.json()
+    return payload.get("documents", []), payload.get("continuation")
+
+
+def _update_document_id_in_vespa(
+    *,
+    index_name: str,
+    old_doc_id: str,
+    new_doc_id: str,
+    user_project_ids: list[int] | None = None,
+) -> None:
+    clean_new_doc_id = replace_invalid_doc_id_characters(new_doc_id)
+    normalized_old = _normalize_legacy_user_file_doc_id(old_doc_id)
+    clean_old_doc_id = replace_invalid_doc_id_characters(normalized_old)
+
+    selection = f"{index_name}.document_id=='{clean_old_doc_id}'"
+    task_logger.debug(f"Vespa selection: {selection}")
+
+    with get_vespa_http_client() as http_client:
+        continuation: str | None = None
+        while True:
+            docs, continuation = _visit_chunks(
+                http_client=http_client,
+                index_name=index_name,
+                selection=selection,
+                continuation=continuation,
+            )
+            if not docs:
+                break
+            for doc in docs:
+                vespa_full_id = doc.get("id")
+                if not vespa_full_id:
+                    continue
+                vespa_doc_uuid = vespa_full_id.split("::")[-1]
+                vespa_url = f"{DOCUMENT_ID_ENDPOINT.format(index_name=index_name)}/{vespa_doc_uuid}"
+                update_request = {
+                    "fields": {"document_id": {"assign": clean_new_doc_id}}
+                }
+                if user_project_ids is not None:
+                    update_request["fields"][USER_PROJECT] = {
+                        "assign": user_project_ids
+                    }
+                r = http_client.put(vespa_url, json=update_request)
+                r.raise_for_status()
+            if not continuation:
+                break
+
+
+@shared_task(
+    name=OnyxCeleryTask.USER_FILE_DOCID_MIGRATION,
+    ignore_result=True,
+    soft_time_limit=LIGHT_SOFT_TIME_LIMIT,
+    time_limit=LIGHT_TIME_LIMIT,
+    bind=True,
+)
+def user_file_docid_migration_task(self: Task, *, tenant_id: str) -> bool:
+    """Per-tenant job to update Vespa and search_doc document_id values for user files.
+
+    - For each user_file with a legacy document_id, set Vespa `document_id` to the UUID `user_file.id`.
+    - Update `search_doc.document_id` to the same UUID string.
+    """
+
+    try:
+        with get_session_with_current_tenant() as db_session:
+            active_settings = get_active_search_settings(db_session)
+            document_index = get_default_document_index(
+                active_settings.primary,
+                active_settings.secondary,
+            )
+            if hasattr(document_index, "index_name"):
+                index_name = document_index.index_name
+            else:
+                index_name = "danswer_index"
+
+            # Fetch mappings of legacy -> new ids
+            rows = db_session.execute(
+                sa.text(
+                    """
+                    SELECT document_id, id
+                    FROM user_file
+                    WHERE document_id IS NOT NULL AND document_id_migrated IS NOT TRUE
+                    """
+                )
+            ).fetchall()
+
+            # dedupe by old document_id
+            seen: set[str] = set()
+            for row in rows:
+                old_doc_id = str(row.document_id)
+                new_uuid = str(row.id)
+                if not old_doc_id or not new_uuid or old_doc_id in seen:
+                    continue
+                seen.add(old_doc_id)
+                # collect user project ids for a combined Vespa update
+                user_project_ids: list[int] | None = None
+                try:
+                    uf = db_session.get(UserFile, UUID(new_uuid))
+                    if uf is not None:
+                        user_project_ids = [project.id for project in uf.projects]
+                except Exception as e:
+                    task_logger.warning(
+                        f"Tenant={tenant_id} failed fetching projects for doc_id={new_uuid}: {e}"
+                    )
+                try:
+                    _update_document_id_in_vespa(
+                        index_name=index_name,
+                        old_doc_id=old_doc_id,
+                        new_doc_id=new_uuid,
+                        user_project_ids=user_project_ids,
+                    )
+                except Exception as e:
+                    task_logger.warning(
+                        f"Tenant={tenant_id} failed Vespa update for {old_doc_id} -> {new_uuid}: {e}"
+                    )
+
+            # Update search_doc records to refer to the UUID string
+            db_session.execute(
+                sa.text(
+                    """
+                    UPDATE search_doc sd
+                    SET document_id = uf.id::text
+                    FROM user_file uf
+                    WHERE uf.document_id IS NOT NULL AND uf.document_id_migrated IS NOT TRUE
+                      AND sd.document_id = uf.document_id
+                    """
+                )
+            )
+            # Mark all processed user_files as migrated
+            db_session.execute(
+                sa.text(
+                    """
+                    UPDATE user_file
+                    SET document_id_migrated = TRUE
+                    WHERE document_id IS NOT NULL AND document_id_migrated IS NOT TRUE
+                    """
+                )
+            )
+            db_session.commit()
+
+        task_logger.info(
+            f"user_file_docid_migration_task completed for tenant={tenant_id} (rows={len(rows)})"
+        )
+        return True
+    except Exception:
+        task_logger.exception(
+            f"user_file_docid_migration_task encountered an error for tenant={tenant_id}"
+        )
+        return False
